+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/index.ts
@@ -14,5 +14,6 @@ export default function apmApiIntegrationTests({
     loadTestFile(require.resolve('./agent_explorer'));
     loadTestFile(require.resolve('./custom_dashboards'));
     loadTestFile(require.resolve('./dependencies'));
+    loadTestFile(require.resolve('./cold_start'));
   });
 }
